Signature into separate file, added optimized programs, not yet used
